Fix bug regarding addition of an already existing .csv file
The application comes with a bunch of pre-existing folders and
.csv files/lists within them. When a user tries to add the
same pre-existing file from the cache the application empties
the (contents of the) csv file.
The bug is fixed by adding a new function in the class
ListActivity.java which checks if the file already exists
when the user tries to add the same file again. So, in
the onActivityResult method, when the file extension is being
checked there is now an additional check for the same file
by calling the new function. If it is indeed the same file,
an appropriate toast message is shown and fasle is returned to
prevent the bug.